decouple msg and events
