+++ examples/demo/main_darwin.go
@@ -63,19 +63,21 @@ func main() {
 	}
 }
 
-func nain() {
-	app.Import(
-		&NavPane{},
-		&Hello{},
-		&Window{},
-	)
+// func nain() {
+// 	app.Import(
+// 		&NavPane{},
+// 		&Hello{},
+// 		&Window{},
+// 	)
 
-	defer app.NewSubscriber().
-		Subscribe("files-openened", nil).
-		Close()
+// 	defer app.NewSubscriber().
+// 		Subscribe("files-openened", nil).
+// 		Close()
 
-	app.Run(
-		mac.Config{},
-		web.Config{},
-	)
-}
+// 	app.Run(
+// 		mac.Config{},
+// 		web.Config{},
+// 	)
+
+// 	// filter right config from target.
+// }
