+++ html.go
@@ -1,5 +1,7 @@
 package app
 
+import "github.com/murlokswarm/app/key"
+
 // HTMLConfig represents a configuration for an html document.
 type HTMLConfig struct {
 	// The title.
@@ -46,3 +48,65 @@ const (
 	DefaultStyleMeta MetaHTTPEquiv = "default-style"
 	RefreshMeta      MetaHTTPEquiv = "refresh"
 )
+
+// MouseEvent represents an onmouse event arg.
+type MouseEvent struct {
+	ClientX   float64
+	ClientY   float64
+	PageX     float64
+	PageY     float64
+	ScreenX   float64
+	ScreenY   float64
+	Button    int
+	Detail    int
+	AltKey    bool
+	CtrlKey   bool
+	MetaKey   bool
+	ShiftKey  bool
+	InnerText string
+	Source    EventSource
+}
+
+// WheelEvent represents an onwheel event arg.
+type WheelEvent struct {
+	DeltaX    float64
+	DeltaY    float64
+	DeltaZ    float64
+	DeltaMode DeltaMode
+	Source    EventSource
+}
+
+// DeltaMode is an indication of the units of measurement for a delta value.
+type DeltaMode uint64
+
+// KeyboardEvent represents an onkey event arg.
+type KeyboardEvent struct {
+	CharCode  rune
+	KeyCode   key.Code
+	Location  key.Location
+	AltKey    bool
+	CtrlKey   bool
+	MetaKey   bool
+	ShiftKey  bool
+	InnerText string
+	Source    EventSource
+}
+
+// DragAndDropEvent represents an ondrop event arg.
+type DragAndDropEvent struct {
+	Files         []string
+	Data          string
+	DropEffect    string
+	EffectAllowed string
+	Source        EventSource
+}
+
+// EventSource represents a descriptor to an event source.
+type EventSource struct {
+	GoappID string
+	CompoID string
+	ID      string
+	Class   string
+	Data    map[string]string
+	Value   string
+}
